Refactor redux folder structure regarding reviewer proposal (move everything to a specific folder, etc.)
