--- src/components/burger-constructor/burger-constructor.slice.ts
@@ -1,174 +1,0 @@
-import { createSelector, PayloadAction } from '@reduxjs/toolkit'
-import produce from 'immer'
-import { memoize } from 'lodash'
-import { nanoid } from 'nanoid'
-import { AnyAction } from 'redux'
-
-import { IBurgerIngredient, IBurgerIngredientUnique } from '../../common/models/data.model'
-import { AppThunk, RootState } from '../../store'
-
-import { IBurgerConstructorIngredientState } from './burger-constructor.model'
-import calculateTotalPrice from './burger-constructor.utils'
-
-/** Actions */
-
-enum ActionKind {
-  AddBun = 'burgerConstructor/addBun',
-  ReplaceBun = 'burgerConstructor/replaceBun',
-  AddTopping = 'burgerConstructor/addTopping',
-  RemoveTopping = 'burgerConstructor/removeTopping',
-  SwapTopping = 'burgerConstructor/swapTopping',
-}
-
-/** Initial state */
-
-export const initialState: IBurgerConstructorIngredientState = {
-  buns: [],
-  toppings: [],
-}
-
-/** Selectors */
-
-export const burgerConstructorSelector = (state: RootState): IBurgerConstructorIngredientState =>
-  state.burgerConstructor
-
-export const bunsSelector = (state: RootState): IBurgerIngredientUnique[] =>
-  state.burgerConstructor.buns
-
-export const toppingsSelector = (state: RootState): IBurgerIngredientUnique[] =>
-  state.burgerConstructor.toppings
-
-export const selectIngredientsTotalPrice = createSelector(
-  [burgerConstructorSelector],
-  (constructor) =>
-    memoize(() => calculateTotalPrice([...constructor.buns, ...constructor.toppings])),
-)
-
-export const selectIngredientsID = createSelector([burgerConstructorSelector], (constructor) =>
-  memoize(() => [...constructor.buns, ...constructor.toppings].map((ingredient) => ingredient._id)),
-)
-
-export const selectIngredientIterationByID = createSelector(
-  [burgerConstructorSelector],
-  (constructor) =>
-    memoize((id: string) =>
-      [...constructor.buns, ...constructor.toppings].reduce((countValue, ingredient) => {
-        if (ingredient._id === id) {
-          return countValue + 1
-        }
-
-        return countValue
-      }, 0),
-    ),
-)
-
-/** Action creators */
-
-const getUniqueIngredientPayload = (ingredient: IBurgerIngredient) => ({
-  ingredient: {
-    ...ingredient,
-    nanoid: nanoid(),
-  },
-})
-
-const addBun = (
-  ingredient: IBurgerIngredient,
-): PayloadAction<{ ingredient: IBurgerIngredientUnique }, ActionKind.AddBun> => ({
-  type: ActionKind.AddBun,
-  payload: getUniqueIngredientPayload(ingredient),
-})
-
-const replaceBun = (
-  ingredient: IBurgerIngredient,
-): PayloadAction<{ ingredient: IBurgerIngredientUnique }, ActionKind.ReplaceBun> => ({
-  type: ActionKind.ReplaceBun,
-  payload: getUniqueIngredientPayload(ingredient),
-})
-
-export const addBunWithReplacement =
-  (ingredient: IBurgerIngredient): AppThunk =>
-  (dispatch, getState) => {
-    const {
-      burgerConstructor: { buns },
-    } = getState()
-
-    if (buns.length && buns[0]._id !== ingredient._id) {
-      dispatch(replaceBun(ingredient))
-    }
-
-    if (!buns.length) {
-      dispatch(addBun(ingredient))
-    }
-  }
-
-export const addTopping = (
-  ingredient: IBurgerIngredient,
-): PayloadAction<{ ingredient: IBurgerIngredientUnique }, ActionKind.AddTopping> => ({
-  type: ActionKind.AddTopping,
-  payload: getUniqueIngredientPayload(ingredient),
-})
-
-export const removeTopping = (
-  ingredient: IBurgerIngredientUnique,
-): PayloadAction<{ ingredient: IBurgerIngredientUnique }, ActionKind.RemoveTopping> => ({
-  type: ActionKind.RemoveTopping,
-  payload: {
-    ingredient,
-  },
-})
-
-export const swapTopping = (
-  toIndex: number,
-  fromIndex: number,
-): PayloadAction<{ toIndex: number; fromIndex: number }, ActionKind.SwapTopping> => ({
-  type: ActionKind.SwapTopping,
-  payload: {
-    toIndex,
-    fromIndex,
-  },
-})
-
-/** Reducer */
-
-export const burgerConstructorReducer = produce((draft, action: AnyAction) => {
-  const { type, payload } = action as PayloadAction<{
-    ingredient: IBurgerIngredientUnique
-    toIndex: number
-    fromIndex: number
-  }>
-
-  switch (type) {
-    case ActionKind.AddBun: {
-      draft.buns.push(payload.ingredient)
-
-      return draft
-    }
-    case ActionKind.ReplaceBun: {
-      draft.buns.pop()
-      draft.buns.push(payload.ingredient)
-
-      return draft
-    }
-    case ActionKind.AddTopping: {
-      draft.toppings.push(payload.ingredient)
-
-      return draft
-    }
-    case ActionKind.RemoveTopping: {
-      const index = draft.toppings.findIndex(
-        (topping) => topping.nanoid === payload.ingredient.nanoid,
-      )
-
-      if (index !== -1) draft.toppings.splice(index, 1)
-
-      return draft
-    }
-    case ActionKind.SwapTopping: {
-      draft.toppings.splice(payload.toIndex, 0, draft.toppings.splice(payload.fromIndex, 1)[0])
-
-      return draft
-    }
-    default:
-      return draft
-  }
-}, initialState)
