+++ src/services/actions/burger-ingredients.actions.ts
@@ -1,53 +1,20 @@
-/* eslint-disable no-param-reassign */
-
-import { AnyAction, createSelector, PayloadAction } from '@reduxjs/toolkit'
-import produce from 'immer'
-import { memoize } from 'lodash'
+import { PayloadAction } from '@reduxjs/toolkit'
 
 import {
   IBurgerIngredient,
   IBurgerIngredientFetch,
-  TBurgerIngredientType,
-  TFetchProcess,
   TFetchProcessError,
   TFetchProcessLoaded,
   TFetchProcessLoading,
 } from '../../common/models/data.model'
-import { AppThunk, RootState } from '../../store'
-import { INGREDIENTS_API_ENDPOINT } from '../app/app.constant'
+import { INGREDIENTS_API_ENDPOINT } from '../../components/app/app.constant'
+import { AppThunk } from '../../store'
 
-import { IBurgerIngredientsState } from './burger-ingredients.model'
-
-/** Actions */
-
-enum ActionKind {
+export enum ActionKind {
   Pending = 'burgerIngredients/requestStatus/pending',
   Fulfilled = 'burgerIngredients/requestStatus/fulfilled',
   Rejected = 'burgerIngredients/requestStatus/rejected',
 }
-
-/** Initial state */
-
-const initialState: IBurgerIngredientsState = {
-  status: 'idle',
-  items: [],
-}
-
-/** Selectors */
-
-export const ingredientsSelector = (state: RootState): IBurgerIngredient[] =>
-  state.burgerIngredients.items
-
-export const ingredientsFetchStatusSelector = (state: RootState): TFetchProcess =>
-  state.burgerIngredients.status
-
-export const selectIngredientsByType = createSelector([ingredientsSelector], (ingredients) =>
-  memoize((type: TBurgerIngredientType) =>
-    ingredients.filter((ingredient) => ingredient.type === type),
-  ),
-)
-
-/** Action creators */
 
 export const startIngredientsFetching = (): PayloadAction<
   { status: TFetchProcessLoading },
@@ -105,34 +72,3 @@ export const fetchIngredients = (): AppThunk => async (dispatch, getState) => {
     }
   }
 }
-
-/** Reducer */
-
-export const burgerIngredientsReducer = produce((draft, action: AnyAction) => {
-  const { type, payload } = action as PayloadAction<{
-    status: TFetchProcess
-    items: IBurgerIngredient[]
-  }>
-
-  switch (type) {
-    case ActionKind.Pending: {
-      draft.status = payload.status
-
-      return draft
-    }
-    case ActionKind.Fulfilled: {
-      draft.items = payload.items
-      draft.status = payload.status
-
-      return draft
-    }
-    case ActionKind.Rejected: {
-      draft.items = []
-      draft.status = payload.status
-
-      return draft
-    }
-    default:
-      return draft
-  }
-}, initialState)
