+++ src/services/selectors/ingredient-details.selector.ts
@@ -1,0 +1,7 @@
+import { IBurgerIngredient } from '../../common/models/data.model'
+import { RootState } from '../../store'
+
+const ingredientDetailsSelector = (state: RootState): IBurgerIngredient | null =>
+  state.ingredientDetails.ingredient
+
+export default ingredientDetailsSelector
